+++ CHANGELOG.md
@@ -75,6 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where colors could appear differently in palette editor to how they appear in game
 - Fix issue where changing one color in a palette could also slightly change the other colors
 - Fix issue where palette editor sliders were not updating after undo and redo
+- Fix issue where "Play Sound Effect" event was using deprecated wait function
 
 ## [4.2.2] - 2026-02-24